requirements.txt
asgiref==3.2.7
Django==3.0.6
django-rest-api==0.1.5
djangorestframework==3.11.0
Markdown==3.2.2
pytz==2020.1
sqlparse==0.3.1
